Computer_Vision/Tensorflow/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.6.2
charset-normalizer==3.3.2
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opencv-python-headless==4.10.0.82  # Using headless version for Cloud Functions
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
protobuf==4.25.3
Pygments==2.18.0
requests==2.32.3
rich==13.7.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1  # Removed tensorflow-intel to avoid conflicts
tensorflow-io-gcs-filesystem
termcolor==2.4.0
typing_extensions==4.12.2
urllib3==2.2.1
wrapt==1.16.0
Pillow==9.2.0
tqdm==4.64.1
Flask==2.3.3
Werkzeug==2.3.7
google-cloud-storage==2.10.0
